module: named anonymous function for debugging purposes
My first contribution

This commit is to help in the effort to name all anonymous
functions to help when heap debugging. The issue asked for
any functions to be updated that are not on a prototype
and this file contains one function that meets those needs.

A previous pull request (13849) was not completed and did
not meet the requirements of the issue so this PR looks
to complete that.

Refs: https://github.com/nodejs/node/issues/8913